backend/requirements.txt
Django==5.2.1
djangorestframework==3.16.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.5.0
Pillow==11.2.1
django-filter==25.1
qrcode==8.2
reportlab==4.4.5
celery==5.6.3
django-celery-beat==2.9.0
redis==7.4.0
django-redis==6.0.0
channels==4.3.2
channels-redis==4.3.0
daphne==4.2.1
